Update the OfferedServices table in DB (autoincrement Id)

+++ CraftsmanContact/Migrations/20240229091251_InitialCreate.cs
@@ -14,7 +14,7 @@ namespace CraftsmanContact.Migrations
                 name: "OfferedServices",
                 columns: table => new
                 {
-                    Id = table.Column<long>(type: "bigint", nullable: false)
+                    Id = table.Column<int>(type: "int", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     Name = table.Column<string>(type: "nvarchar(max)", nullable: false),
                     Description = table.Column<string>(type: "nvarchar(max)", nullable: true)
