+++ CraftsmanContact/Migrations/OfferedServiceContextModelSnapshot.cs
@@ -23,11 +23,11 @@ namespace CraftsmanContact.Migrations
 
             modelBuilder.Entity("CraftsmanContact.Models.OfferedService", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
                     b.Property<string>("Description")
                         .HasColumnType("nvarchar(max)");
